requirements.txt
markdownify
python-docx
markdown2
requests
deep-translator
beautifulsoup4
lxml
